gap: Fix mouse pos on win32 non-primary monitor

+++ libs/gap/src/pal_win32.c
@@ -294,7 +294,7 @@ pal_event(GapPal* pal, const HWND wnd, const UINT msg, const WPARAM wParam, cons
     ValidateRect(wnd, null);
     return true;
   case WM_MOUSEMOVE: {
-    const GapVector newPos = gap_vector(LOWORD(lParam), HIWORD(lParam));
+    const GapVector newPos = gap_vector((i32)(short)LOWORD(lParam), (i32)(short)HIWORD(lParam));
     pal_event_cursor(window, newPos);
     return true;
   }
